% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnCheckInputsMelding.R
\name{fnCheckInputsMelding}
\alias{fnCheckInputsMelding}
\title{This is a function to check the inputs of fnPredictMelding()
check they are sf objects
check CRS points and areas are the same.}
\usage{
fnCheckInputsMelding(depoint, dearea, dppoint, dparea, boundaryregion)
}
\arguments{
\item{depoint}{dataset estimation point.sf obj. Columns value, geometry}

\item{dearea}{dataset estimation area. sf obj. Columns value, geometry}

\item{dppoint}{dataset prediction point. sf obj. Columns geometry}

\item{dparea}{dataset prediction area. sf obj. Columns geometry}

\item{boundaryregion}{The boundary region of the mesh. This is a sf object. The input data and the target should be within the boundary.}
}
\value{
if the inputs of fn
}
\description{
This is a function to check the inputs of fnPredictMelding()
check they are sf objects
check CRS points and areas are the same.
}
