% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnPredictMelding.R
\name{fnPredictMelding}
\alias{fnPredictMelding}
\title{This is the function to predicts using the melding approach.}
\usage{
fnPredictMelding(
  depoint = NULL,
  dearea = NULL,
  dppoint = NULL,
  dparea = NULL,
  boundaryregion,
  mesh = NULL,
  priorspdesigma = NULL,
  priorspderange = NULL
)
}
\arguments{
\item{depoint}{dataset estimation point.sf obj. Columns value, geometry}

\item{dearea}{dataset estimation area. sf obj. Columns value, geometry}

\item{dppoint}{dataset prediction point. sf obj. Columns geometry}

\item{dparea}{dataset prediction area. sf obj. Columns geometry}

\item{boundaryregion}{The boundary region of the mesh. This is a sf object. The input data and the target should be within the boundary.}

\item{mesh}{obj}

\item{priorspdesigma}{a positive two-length vector. Construct SPDE with pc prior}

\item{priorspderange}{a positive two-length vector. Construct SPDE with pc prior}
}
\value{
a sf object including prediction results.
}
\description{
This is the function to predicts using the melding approach.
}
