% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialm.R
\docType{package}
\name{spatialm}
\alias{spatialm}
\title{spatialm: A package for predicting the spatial misaligned data using SPDE}
\description{
The spatialm package provides two categories of important functions:
fnCreatMesh, fnPredictMelding.
}
\section{spatialm functions}{

The fnCreatMesh functions helps users to create a fine mesh.

The fnPredictMelding use Bayesian Melding Methdos provide by Moraga (2017)
}

